chart-app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
altair==6.0.0
    # via -r requirements.in
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   starlette
    #   watchfiles
async-lru==2.1.0
    # via -r requirements.in
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
fastapi[all]==0.128.0
    # via -r requirements.in
fastapi-cli[standard]==0.0.20
    # via fastapi
fastapi-cloud-cli==0.11.0
    # via fastapi-cli
fastar==0.8.0
    # via fastapi-cloud-cli
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.6
    # via
    #   altair
    #   fastapi
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
jupyterhealth-client==0.1.0a3
    # via -r requirements.in
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.7.1
    # via yarl
narwhals==2.15.0
    # via altair
numpy==2.4.1
    # via pandas
oauthlib==3.3.1
    # via requests-oauthlib
orjson==3.11.6
    # via fastapi
packaging==26.0
    # via altair
pandas==3.0.0
    # via
    #   -r requirements.in
    #   jupyterhealth-client
propcache==0.4.1
    # via yarl
pydantic[email]==2.12.5
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-extra-types==2.11.0
    # via fastapi
pydantic-settings==2.12.0
    # via
    #   -r requirements.in
    #   fastapi
pygments==2.19.2
    # via rich
pyjwt==2.11.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.22
    # via fastapi
pyyaml==6.0.3
    # via
    #   fastapi
    #   uvicorn
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   jupyterhealth-client
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via -r requirements.in
rich==14.3.1
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.17.2
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.7.6
    # via fastapi-cloud-cli
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
sentry-sdk==2.51.0
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
starlette==0.50.0
    # via fastapi
typer==0.21.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.15.0
    # via
    #   altair
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   rich-toolkit
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
ujson==5.11.0
    # via fastapi
urllib3==2.6.3
    # via
    #   requests
    #   sentry-sdk
uvicorn[standard]==0.40.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
yarl==1.22.0
    # via jupyterhealth-client
